ストリーミングから流れてきた通知もキャッシュするようにした

+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/notification/impl/NotificationPagingStoreImpl.kt
@@ -82,6 +82,9 @@ class NotificationPagingStoreImpl(
     }
 
     override suspend fun loadPrevious(): Result<Unit> = runCancellableCatching {
+        if (delegate.mutex.isLocked) {
+            return@runCancellableCatching
+        }
         unreadNotificationDAO.deleteWhereAccountId(getAccount().accountId)
         previousPagingController.loadPrevious().getOrThrow()
     }
